Update registration process and restrict user permissions
This commit updates the registration process to include password confirmation and restricts access to creating new recipes to admin users only. The registration page on the front-end has been updated to allow for password confirmation. Also, the main routes for viewing all recipes and a specific recipe no longer require login, and the user model now includes an 'admin' attribute.